requirements.txt
Flask==3.0.3
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
APScheduler==3.10.4
Flask-Caching==2.3.0
Werkzeug==3.0.1
